% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{causetextV5}
\alias{causetextV5}
\title{Translation list of COD codes}
\format{
A data frame with the translation of codes to their names for 3
pregnancy statuses, 61 CODs (both the version of COD only and COD with group code),
and 6 circumstances of mortality (COMCAT).
}
\description{
This is the translation of COD abbreviation codes into their corresponding
full names.
}
\examples{

data(causetextV5)

}
\keyword{datasets}
