% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{package}
\name{InterVA5-package}
\alias{InterVA5-package}
\title{Perform InterVA5 algorithm and provide graphical summarization of COD
distribution.}
\description{
Computes individual cause of death and population cause-specific mortality
fractions using the InterVA5 algorithm. Provides a simple graphical
representation of the result.
}
\details{
To get the most up-to-date version of the package, as well as the past
versions, please check the github repository at:
\url{https://github.com/jarathomas/InterVA5-R-Replicate/}

\tabular{ll}{ Package: \tab InterVA5\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2018-02-01\cr License: \tab GPL-3\cr }
}
\examples{

data(SampleInputV5)
sample.output <- InterVA5(SampleInputV5, HIV = "h", Malaria = "v", 
    write = TRUE, directory = tempdir(), filename = "VA5_result", 
    output = "extended", append = FALSE)

}
\references{
http://www.interva.net/
}
\author{
Jason Thomas, Zehang Li, Tyler McCormick, Sam Clark

Maintainer: Jason Thomas <jarathomas@gmail.com>
}
\keyword{InterVA}
