% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5.summary.R
\name{print.interVA5_summary}
\alias{print.interVA5_summary}
\title{Print method for summary of the results obtained from InterVA5 algorithm}
\usage{
\method{print}{interVA5_summary}(x, ...)
}
\arguments{
\item{x}{summary of InterVA5 results}

\item{...}{not used}
}
\description{
This function prints the summary message of the fitted results.
}
\keyword{InterVA}
