% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{probbase}
\alias{probbase}
\title{Conditional probability of InterVA5}
\format{A data frame with 354 observations on 87 variables. Each observation
is the conditional probability.}
\description{
This is the table of conditional probabilities of symptoms given CODs. The
values are from InterVA-5
}
\examples{

data(probbase)

}
\keyword{datasets}
