% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{probbaseV5}
\alias{probbaseV5}
\title{Conditional probability of InterVA5}
\format{A data frame with 354 observations on 87 variables. The first row contains
observations corresponding to prior probabilities; while the subsequent observations
(rows 2 - 354) are the conditional probabilities.}
\description{
This is the table of conditional probabilities of symptoms given CODs, along with
prior probabilities in the first row. The
values are from InterVA-5
}
\examples{

data(probbaseV5)

}
\keyword{datasets}
