% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{probbaseV5_14}
\alias{probbaseV5_14}
\title{Version 14 of the conditional probability of InterVA5}
\format{A data frame with 354 observations on 87 variables. The first row contains
observations corresponding to prior probabilities; while the subsequent observations
(rows 2 - 354) are the conditional probabilities.}
\description{
This is version 14 (February 15th, 2018) of the table of conditional probabilities of symptoms given CODs, along with
prior probabilities in the first row. The
values are from InterVA-5
}
\examples{

data(probbaseV5_14)

}
\keyword{datasets}
