% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{probbaseV5_19}
\alias{probbaseV5_19}
\title{Version 19 of the conditional probability of InterVA5}
\format{
A data frame with 354 observations on 87 variables. The first row contains
observations corresponding to prior probabilities; while the subsequent observations
(rows 2 - 354) are the conditional probabilities.
}
\description{
This is version 19 (July 20, 2021) of the table of conditional probabilities of symptoms given CODs, along with
prior probabilities in the first row. The
values differ from the last version (v18) of InterVA-5 (interva.net) by setting
Pr(abortion-related death | i309 = 1) = "N"
Pr(abortion-related death | i310 = 1) = "N"
(the previous values were "E").
}
\examples{

data(probbaseV5_19)

}
\keyword{datasets}
