% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{SampleInputV5}
\alias{SampleInputV5}
\title{10 records of Sample Input}
\format{10 arbitrary input records.}
\description{
This is a dataset consisting of 10 arbitrary sample input deaths in the
acceptable format of InterVA5. The last row in this dataset only has missing values,
as indicated in the error log generated by InterVA5.  Any dataset that needs to be analyzed by this
package should be in the same format. The order of the input fields must
not be changed.
}
\examples{

data(SampleInputV5)

}
\keyword{datasets}
