% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateProbbase.R
\name{update.SCI}
\alias{update.SCI}
\title{Update the Symptom-Cause-Information source (aka probbaseV5).}
\usage{
\method{update}{SCI}()
}
\value{
\item{newProbbase}{The Symptom-Cause-Information ("Probbase") used to assign causes of death.}
}
\description{
The function takes input of a list of va object and calculates the
cause-specific mortality fraction. It only calculates CSMF5 as aggregation of up to the third largest causes.
}
\examples{

\dontrun{
data(RandomVA5)
RandomVA5 <- RandomVA5[1:2, ]
newProbbase <- update.SCI()
out <- InterVA5(RandomVA5, sci = newProbbase, HIV = "h", Malaria = "l", write=FALSE, 
    directory = tempdir(), filename = "VA5_result", output = "extended", append = FALSE)
}

}
\seealso{
\code{\link{probbaseV5}}
}
\author{
Jason Thomas, Zehang LI, Tyler McCormick, Sam Clark
}
\keyword{interVA}
