% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadProbbase.R
\name{version.SCI}
\alias{version.SCI}
\title{Print the version of the Symptom-Cause-Information source (aka probbaseV5).}
\usage{
version.SCI(sci)
}
\arguments{
\item{sci}{a symptom-cause-information matrix}
}
\value{
\item{}{Message stating the Probbase version.}
}
\description{
The function takes
}
\examples{

\dontrun{
data(probbaseV5)
version.SCI(sci = probbaseV5)
}

}
\seealso{
\code{\link{download.SCI}}
}
\author{
Jason Thomas, Zehang LI, Tyler McCormick, Sam Clark
}
\keyword{interVA}
