% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterVA5_1.0-package.R
\docType{data}
\name{SampleInputV5}
\alias{SampleInputV5}
\title{3 records of Sample Input}
\format{3 arbitrary input records.}
\description{
This is a dataset consisting of 3 arbitrary sample input deaths in the
acceptable format of InterVA5. Any dataset that needs to be analyzed by this
package should be in the same format. The order of the input fields must
not be changed.
}
\examples{

data(SampleInputV5)

}
\keyword{datasets}
